requirements.txt
aioredis==1.3.1
appdirs==1.4.3
asgiref==3.2.10
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
certifi==2019.11.28
cffi==1.14.3
channels==2.4.0
channels-redis==2.4.2
colorama==0.4.3
constantly==15.1.0
cryptography==3.1.1
daphne==2.5.0
distlib==0.3.0
dj-database-url==0.5.0
Django==3.1
django-ckeditor==6.0.0
django-cors-headers==3.5.0
django-filter==2.3.0
django-fsm==2.7.1
django-heroku==0.3.1
django-js-asset==1.2.2
django-model-utils==4.0.0
django-simple-history==2.11.0
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
drf-writable-nested==0.6.1
env==0.1.0
filelock==3.0.12
gunicorn==20.0.4
h2==3.2.0
hiredis==1.1.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==20.0.1
idna==2.10
importlib-metadata==1.6.0
importlib-resources==1.4.0
incremental==17.5.0
isort==5.5.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==0.6.2
pipenv==2018.11.26
priority==1.3.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.6.0
pyOpenSSL==19.1.0
python-decouple==3.3
pytz==2020.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
virtualenv==20.0.15
virtualenv-clone==0.5.4
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.1.0
zope.interface==5.1.2

python-dotenv~=0.14.0